requirements.txt
flask>=3.0
python-dotenv>=1.0
google-genai>=1.0.0
googlemaps>=4.10.0
requests>=2.31
